+++ data/src/main/java/gr/tsambala/tutorbilling/data/di/DaoModule.kt
@@ -1,0 +1,19 @@
+package gr.tsambala.tutorbilling.data.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import gr.tsambala.tutorbilling.data.dao.LessonDao
+import gr.tsambala.tutorbilling.data.dao.StudentDao
+import gr.tsambala.tutorbilling.data.database.TutorBillingDatabase
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DaoModule {
+    @Provides
+    fun provideStudentDao(db: TutorBillingDatabase): StudentDao = db.studentDao()
+
+    @Provides
+    fun provideLessonDao(db: TutorBillingDatabase): LessonDao = db.lessonDao()
+}